+++ pytket/docs/conf.py
@@ -38,9 +38,9 @@ copyright = "2023 Quantinuum"
 author = "Quantinuum"
 
 # The short X.Y version
-version = "1.17"
+version = "1.18"
 # The full version, including alpha/beta/rc tags
-release = "1.17.1"
+release = "1.18.0"
 
 
 # -- General configuration ---------------------------------------------------