+++ tket/include/tket/Circuit/ToffoliBox.hpp
@@ -73,10 +73,7 @@ class ToffoliBox : public Box {
   /**
    * Equality check between two ToffoliBox instances
    */
-  bool is_equal(const Op &op_other) const override {
-    const ToffoliBox &other = dynamic_cast<const ToffoliBox &>(op_other);
-    return id_ == other.get_id();
-  }
+  bool is_equal(const Op &op_other) const override;
 
   std::optional<Eigen::MatrixXcd> get_box_unitary() const override;