+++ pytket/docs/changelog.rst
@@ -1,6 +1,35 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Minor new features:
+
+* Implement equality checking for all boxes.
+* Add ``Op.is_clifford`` to python binding.
+* Single-qubit squashing ignores chains of symbolic gates if squashing them
+  would increase the overall complexity of the expressions. This behaviour can
+  be overridden using the ``always_squash_symbols`` parameter to
+  ``SquashCustom``.
+
+Fixes:
+
+* Allow ``BackendResult`` objects containing no results.
+
+1.18.0 (August 2023)
+--------------------
+
+Minor new features:
+
+* Add circuit method ``depth_2q``.
+* Add ``allow_swaps`` parameter to ``auto_rebase_pass``.
+
+Fixes:
+
+* Fix slow ``Circuit.get_statevector()``.
+
+
 1.17.1 (July 2023)
 ------------------